% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/funs.r
\name{look}
\alias{look}
\title{Display lines of a file}
\usage{
look(file, key, keyEnd = NULL, dflag = FALSE, skip = 0L, maxLines = -1L,
  raw = FALSE)
}
\arguments{
\item{file}{name of the file to search within}

\item{key}{a length one character vector; the key
to search for.}

\item{keyEnd}{an optional length one character vector;
if set, results will be given for all rows
with starting characters between key and keyEnd}

\item{dflag}{logical; flag for whether only alphanumeric
characters are being searched over. When set,
the function will run slightly more efficiently.}

\item{skip}{number of rows to skip in the input file; these do
not need to be sorted (i.e., a row of headers or
rows of comments)}

\item{maxLines}{maximum number of lines to return. Set to Inf
or -1 to return all lines.}

\item{raw}{logical. Should results be returned as a raw
vector, or a character vector split by newline
characters.}
}
\description{
This function provides a more direct analogue to the
BSD function 'look' for users looking for more control
on the output. The file must be pre-sorted on the row
prefixes
}

