% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/funs.r
\name{saveAsFileDB}
\alias{saveAsFileDB}
\title{Save a data frame as fileDB file}
\usage{
saveAsFileDB(df, file)
}
\arguments{
\item{df}{data frame to write to disk. First column
will be 'indexed'; must be a character or
factor as numerics will not sort correctly
unless they all have the same number of digits}

\item{file}{name of the file to store the results in.}
}
\description{
This function saves a data frame as a flat file
that can be searched using \code{\link{scan}}
and \code{\link{search}}. The first row of the
output is a commented out string of column types;
the second row are column names and the remainder
is a pipe-deliminated file. Conveniently, this can
also be read back in using \code{\link{read.table}},
by setting sep='|'.
}

