% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/funs.r
\name{scanFile}
\alias{scanFile}
\title{Scan for key in fileDB file}
\usage{
scanFile(file, keyStart, keyEnd, formatter, header)
}
\arguments{
\item{file}{name of the file to search within}

\item{keyStart}{a length one character vector; the key
to start searching for.}

\item{keyEnd}{a length one character vector; the key
to start searching for. Results will be
given for all rows with starting characters
between keyStart and keyEnd}

\item{formatter}{function which accepts a raw vector and
returns the results. If missing, this will
by constructed via the file metadata.}

\item{header}{logical. Whether file has a head of information,
as supplied by the function saveAsFileDB. When
missing, will be infered from the first line of
the file.}
}
\description{
This function provides a means for scan a sorted
flat file constructed by \code{\link{saveAsFileDB}},
returning all rows with prefixes between two keys.
}

