% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/funs.r
\name{search}
\alias{search}
\title{Search for key in fileDB file}
\usage{
search(file, key, formatter)
}
\arguments{
\item{file}{name of the file to search within}

\item{key}{a length one character vector; the key
to search for.}

\item{formatter}{function which accepts a raw vector and
returns the results. If missing, this will
by constructed via the file metadata.}
}
\description{
This function provides a means for searching a sorted
flat file constructed by \code{\link{saveAsFileDB}}.
Use \code{link{scan}} for a range query.
}

