% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/funs.r
\name{searchFile}
\alias{searchFile}
\title{Search for key in fileDB file}
\usage{
searchFile(file, key, formatter, maxLines = 100L, header)
}
\arguments{
\item{file}{name of the file to search within}

\item{key}{a length one character vector; the key
to search for.}

\item{formatter}{function which accepts a raw vector and
returns the results. If missing, this will
by constructed via the file metadata.}

\item{maxLines}{maximum number of lines to return. Set to Inf
or -1 to return all lines.}

\item{header}{logical. Whether file has a head of information,
as supplied by the function saveAsFileDB. When
missing, will be infered from the first line of
the file.}
}
\description{
This function provides a means for searching a sorted
flat file constructed by \code{\link{saveAsFileDB}}.
Use \code{link{scanFile}} for a range query.
}

